+++ src/features/community/apis/post.ts
@@ -5,6 +5,7 @@ import {
   PostDetailServerResp,
   PostsListResp,
   RequestPageParams,
+  UpdatePostBody,
 } from '../types/postsListType';
 
 // 게시글 목록 조회
@@ -30,16 +31,14 @@ export async function createPost(boardId: number, body: CreatePostBody) {
   return data as string;
 }
 
-// 게시글 수정
-// export async function putEditPost(postId: string, edited: {
-//   content: string;
-//   images: string | string[] | null;
-//   removedImages: string | string[] | null;
-// }) {
-//   const response = await api.put(`/api/v1/posts/${postId}`, edited)
-// }
+// 게시글 삭제
+export async function deletePost(postId: number) {
+  await api.delete(`/api/v1/posts/${postId}`);
+  return true;
+}
 
-//게시글 삭제
-// export async function deletePost(postId: string) {
-//   return await api.delete(`/api/v1/posts/${postId}`)
-// }
+// 게시글 수정
+export async function updatePost(postId: number, body: UpdatePostBody) {
+  await api.put(`/api/v1/posts/${postId}`, body);
+  return true;
+}
